requirements.txt
django-debug-toolbar==1.5
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==3.0
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
Django==1.8.6
django-push-notifications==1.3.1
djangorestframework==3.6.2
idna==2.5
oauthlib==2.0.2
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
urllib3==1.21.1
